+++ src/runtime.rs
@@ -156,10 +156,7 @@ pub enum Value {
 
     Int(i64),
 
-    // TODO: Should probably be renamed to Float, and I should just add a
-    // denominator to my existing decimal type if I want to represent fractions
-    // exactly.
-    Decimal(Decimal),
+    Float(Decimal),
 
     String(Rc<str>),
 
@@ -235,11 +232,11 @@ impl Value {
             (Type::Null, Value::Null) => return Ok(()),
             (Type::Bool, Value::Bool(..)) => return Ok(()),
             (Type::Int, Value::Int(..)) => return Ok(()),
-            (Type::Float, Value::Decimal(..)) => return Ok(()),
+            (Type::Float, Value::Float(..)) => return Ok(()),
             (Type::String, Value::String(..)) => return Ok(()),
 
             // Int and Float are numbers.
-            (Type::Num, Value::Int(..) | Value::Decimal(..)) => return Ok(()),
+            (Type::Num, Value::Int(..) | Value::Float(..)) => return Ok(()),
 
             // For compound types, we descend into them to check.
             (Type::List(elem_type), Value::List(elems)) => {
